[C API] Moved the Xamarin code into separate files to avoid conflicts
