+++ gyp/core.gypi
@@ -11,11 +11,50 @@
 #
 {
     'sources': [
+        # Xamarin created files
+        '<(skia_include_path)/c/xamarin/sk_x_types.h',
+        '<(skia_include_path)/c/xamarin/sk_x_bitmap.h',
+        '<(skia_include_path)/c/xamarin/sk_x_imagedecoder.h',
+        '<(skia_include_path)/c/xamarin/sk_x_canvas.h',
+        '<(skia_include_path)/c/xamarin/sk_x_data.h',
+        '<(skia_include_path)/c/xamarin/sk_x_image.h',
+        '<(skia_include_path)/c/xamarin/sk_x_maskfilter.h',
+        '<(skia_include_path)/c/xamarin/sk_x_imagefilter.h',
+        '<(skia_include_path)/c/xamarin/sk_x_colorfilter.h',
+        '<(skia_include_path)/c/xamarin/sk_x_paint.h',
+        '<(skia_include_path)/c/xamarin/sk_x_path.h',
+        '<(skia_include_path)/c/xamarin/sk_x_shader.h',
+        '<(skia_include_path)/c/xamarin/sk_x_stream.h',
+        '<(skia_include_path)/c/xamarin/sk_x_typeface.h',
+        '<(skia_src_path)/c/xamarin/sk_x_types_priv.h',
+        '<(skia_src_path)/c/xamarin/sk_x_bitmap.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_imagedecoder.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_canvas.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_data.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_image.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_maskfilter.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_imagefilter.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_colorfilter.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_paint.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_path.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_shader.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_stream.cpp',
+        '<(skia_src_path)/c/xamarin/sk_x_typeface.cpp',
+        
+        '<(skia_include_path)/c/sk_canvas.h',
+        '<(skia_include_path)/c/sk_data.h',
+        '<(skia_include_path)/c/sk_image.h',
+        '<(skia_include_path)/c/sk_maskfilter.h',
+        '<(skia_include_path)/c/sk_matrix.h',
+        '<(skia_include_path)/c/sk_paint.h',
+        '<(skia_include_path)/c/sk_path.h',
+        '<(skia_include_path)/c/sk_picture.h',
+        '<(skia_include_path)/c/sk_shader.h',
+        '<(skia_include_path)/c/sk_surface.h',
+        '<(skia_include_path)/c/sk_types.h',
+        '<(skia_src_path)/c/sk_c_from_to.h',
         '<(skia_src_path)/c/sk_paint.cpp',
         '<(skia_src_path)/c/sk_surface.cpp',
-        '<(skia_src_path)/c/sk_typeface.cpp',
-        '<(skia_src_path)/c/sk_stream.cpp',
-        '<(skia_src_path)/c/sk_bitmap.cpp',
         '<(skia_src_path)/c/sk_types_priv.h',
 
         '<(skia_src_path)/core/Sk4px.h',
