+++ include/c/sk_canvas.h
@@ -20,26 +20,7 @@ SK_C_PLUS_PLUS_BEGIN_GUARD
     balancing call to sk_canvas_restore() is made, the previous matrix
     and clip are restored.
 */
-SK_API void sk_canvas_clear(sk_canvas_t*);
-/**
-   This makes the contents of the canvas undefined. Subsequent calls that
-   require reading the canvas contents will produce undefined results. Examples
-   include blending and readPixels. The actual implementation is backend-
-   dependent and one legal implementation is to do nothing. Like clear(), this
-   ignores the clip.
-   
-   This function should only be called if the caller intends to subsequently
-   draw to the canvas. The canvas may do real work at discard() time in order
-   to optimize performance on subsequent draws. Thus, if you call this and then
-   never draw to the canvas subsequently you may pay a perfomance penalty.
-*/
-SK_API void sk_canvas_discard(sk_canvas_t*);
-/**
-    Save the current matrix and clip on the canvas.  When the
-    balancing call to sk_canvas_restore() is made, the previous matrix
-    and clip are restored.
-*/
-SK_API int sk_canvas_save(sk_canvas_t*);
+SK_API void sk_canvas_save(sk_canvas_t*);
 /**
     This behaves the same as sk_canvas_save(), but in addition it
     allocates an offscreen surface. All drawing calls are directed
@@ -56,7 +37,7 @@ SK_API int sk_canvas_save(sk_canvas_t*);
                        to the offscreen when sk_canvas_restore() is
                        called.
 */
-SK_API int sk_canvas_save_layer(sk_canvas_t*, const sk_rect_t*, const sk_paint_t*);
+SK_API void sk_canvas_save_layer(sk_canvas_t*, const sk_rect_t*, const sk_paint_t*);
 /**
     This call balances a previous call to sk_canvas_save() or
     sk_canvas_save_layer(), and is used to remove all modifications to
@@ -65,19 +46,6 @@ SK_API int sk_canvas_save_layer(sk_canvas_t*, const sk_rect_t*, const sk_paint_t
     save_layer were called.
 */
 SK_API void sk_canvas_restore(sk_canvas_t*);
-/**
-    Returns the number of matrix/clip states on the SkCanvas' private stack.
-    This will equal # save() calls - # restore() calls + 1. The save count on
-    a new canvas is 1.
-*/
-SK_API int sk_canvas_get_save_count(sk_canvas_t*);
-/**
-    Efficient way to pop any calls to sk_canvas_save() that happened after the save
-    count reached saveCount. It is an error for saveCount to be greater than
-    getSaveCount(). To pop all the way back to the initial matrix/clip context
-    pass saveCount == 1.
-*/
-SK_API void sk_canvas_restore_to_count(sk_canvas_t*, int saveCount);
 
 /**
     Preconcat the current coordinate transformation matrix with the
@@ -128,41 +96,6 @@ SK_API void sk_canvas_clip_path(sk_canvas_t*, const sk_path_t*);
     specified paint.
 */
 SK_API void sk_canvas_draw_paint(sk_canvas_t*, const sk_paint_t*);
-/**
-    Draws with the specified color and mode.
-**/
-SK_API void sk_canvas_draw_color(sk_canvas_t* ccanvas, sk_color_t color, sk_xfermode_mode_t mode);
-/**
-   Draw a series of points, interpreted based on the sk_point_mode_t mode. For
-   all modes, the count parameter is interpreted as the total number of
-   points. For LINES_SK_POINT_MODE mode, count/2 line segments are drawn.
-   For POINTS_SK_POINT_MODE mode, each point is drawn centered at its coordinate, and its
-   size is specified by the paint's stroke-width. It draws as a square,
-   unless the paint's cap-type is round, in which the points are drawn as
-   circles.
-   For LINES_SK_POINT_MODE mode, each pair of points is drawn as a line segment,
-   respecting the paint's settings for cap/join/width.
-   For POLYGON_SK_POINT_MODE mode, the entire array is drawn as a series of connected
-   line segments.
-   Note that, while similar, LINES_SK_POINT_MODE and POLYGON_SK_POINT_MODE modes draw slightly
-   differently than the equivalent path built with a series of moveto,
-   lineto calls, in that the path will draw all of its contours at once,
-   with no interactions if contours intersect each other (think XOR
-   xfermode). sk_canvas_draw_paint always draws each element one at a time.
-*/
-SK_API void sk_canvas_draw_points(sk_canvas_t*, sk_point_mode_t, size_t, const sk_point_t[], const sk_paint_t*);
-/**
-   Draws a single point with the specified paint
-*/
-SK_API void sk_canvas_draw_point(sk_canvas_t*, float, float, const sk_paint_t*);
-/**
-   Draws a single point with the specified paint
-*/
-SK_API void sk_canvas_draw_point_color(sk_canvas_t*, float, float, sk_color_t);
-/**
-   Draws a line from x0,y0 to x1,y1
-*/
-SK_API void sk_canvas_draw_line(sk_canvas_t* ccanvas, float x0, float y0, float x1, float y1, sk_paint_t* cpaint);
 /**
     Draw the specified rectangle using the specified paint. The
     rectangle will be filled or stroked based on the style in the
@@ -215,39 +148,6 @@ SK_API void sk_canvas_draw_image_rect(sk_canvas_t*, const sk_image_t*,
 */
 SK_API void sk_canvas_draw_picture(sk_canvas_t*, const sk_picture_t*,
                                    const sk_matrix_t*, const sk_paint_t*);
-
-/**
-   Draw the text, with origin at (x,y), using the specified paint.
-   The origin is interpreted based on the Align setting in the paint.
-	
-   @param text The text to be drawn
-   @param byteLength   The number of bytes to read from the text parameter
-   @param x        The x-coordinate of the origin of the text being drawn
-   @param y        The y-coordinate of the origin of the text being drawn
-   @param paint    The paint used for the text (e.g. color, size, style)
-*/
-SK_API void sk_canvas_draw_text (sk_canvas_t*, const char *text, size_t byteLength, float x, float y, const sk_paint_t* paint);
-/**
-    Draw the text, with each character/glyph origin specified by the pos[]
-    array. The origin is interpreted by the Align setting in the paint.
-    @param text The text to be drawn
-    @param byteLength   The number of bytes to read from the text parameter
-    @param pos      Array of positions, used to position each character
-    @param paint    The paint used for the text (e.g. color, size, style)
-*/
-SK_API void sk_canvas_draw_pos_text (sk_canvas_t*, const char *text, size_t byteLength, const sk_point_t[], const sk_paint_t* paint);
-
-SK_API void sk_canvas_draw_text_on_path (sk_canvas_t*, const char *text, size_t byteLength, const sk_path_t*path, float hOffset, float vOffset, const sk_paint_t* paint);
-
-SK_API void sk_canvas_draw_bitmap(sk_canvas_t* ccanvas, const sk_bitmap_t* cbitmap, float x, float y, const sk_paint_t* cpaint);
-
-SK_API void sk_canvas_draw_bitmap_rect(sk_canvas_t* ccanvas, const sk_bitmap_t* cbitmap, const sk_rect_t* csrcR, const sk_rect_t* cdstR, const sk_paint_t* cpaint);
-
-SK_API void sk_canvas_reset_matrix(sk_canvas_t* ccanvas);
-
-SK_API void sk_canvas_set_matrix(sk_canvas_t* ccanvas, const sk_matrix_t* cmatrix);
-
-SK_API void sk_canvas_get_total_matrix(sk_canvas_t* ccanvas, sk_matrix_t* cmatrix);
 
 SK_C_PLUS_PLUS_END_GUARD
