+++ include/effects/SkDropShadowImageFilter.h
@@ -5,12 +5,12 @@
  * found in the LICENSE file.
  */
 
+#ifndef SkDropShadowImageFilter_DEFINED
+#define SkDropShadowImageFilter_DEFINED
+
 #include "SkColor.h"
 #include "SkImageFilter.h"
 #include "SkScalar.h"
-
-#ifndef SkDropShadowImageFilter_DEFINED
-#define SkDropShadowImageFilter_DEFINED
 
 class SK_API SkDropShadowImageFilter : public SkImageFilter {
 public:
